[QA-BitVMX] - Separated tests to a specific folder. Added 4 new test cases. Added test readme.txt
